Proportional contribution of an ancestral area to the species in other region

For the species present in a site, the function calculates the proportional contribution of ancestral areas to dispersal of lineage in the site's region.

dispersal_from(W, tree, ancestral.area, biogeo)

Arguments

W
Occurrence matrix, rows are assemblages and columns are species

tree
Phylogenetic tree in newick format

ancestral.area
One column data frame with nodes in rows and one column indicating the occurrence (biome/ecoregion) area of nodes

biogeo
One column data frame with assemblages in rows and their respective biome/ecoregion

Value

See also

Author

Arthur V Rodrigues rodrigues.arthur.v@gmail.com

Examples

if (FALSE) {
data(akodon.sites) # occurrence matrix
data(akodon.pa.tree) # phylogenetic tree
data(regions) # biogeographic region
data(resDEC) # output from ancestral area reconstruction
node.area <- get_node_range_BioGeoBEARS(resDEC,phyllip.file = here("inst", "extdata", "geo_area_akodon.data"),akodon.newick,max.range.size = 4) # data frame indicating the ancestral area occurrence
dispersal_from(W=akodon.sites,tree=akodon.pa.tree,ancestral.area=node.area,biogeo=regions) # historical dispersal analysis
}