Tip-based metrics of trait evolution

Tip-based metrics of trait evolution

tip.based.trait.evo(
  tree,
  trait,
  nsim = 1,
  method = c("transition_rates", "last_transition_time", "stasis_time")
)

Arguments

tree

A phylogenetic tree as an object of class "phylo"

trait

A named vector containing the tip states for a discretely valued character. The names must match the tip labels of the tree.

nsim

Number of simulations to stochastic character mapping.

method

Tip-based metric, partial match to "transition_rates", "last_transition_time" and "stasis_time".

Value

A list (length equal to nsim) with tip-based metrics estimated per species.

Author

André Luza and Vanderlei Debastiani