Defining biogeographic regions based on phylogenetic turnover

Defining biogeographic regions based on phylogenetic turnover

Usage

evoregions(
  comm,
  phy,
  max.n.clust = NULL,
  method.dist = "bray",
  tresh.dist = 0.05,
  method.clust = "kmeans",
  stat.clust = "BIC",
  n.iter.clust = 1e+07,
  criterion.clust = "diffNgroup"
)

Arguments

comm

Species occurrence matrix. Assemblages are rows and species are columns

phy

phylogenetic tree of class phylo containing the species in comm.

max.n.clust

Integer value to be used in find.clusters. Indicates the maximum number of clusters to be tried.

method.dist

Character. The method to be used to compute phyogenetic distances among assemblages. Dissimilarity index, as accepted by vegdist (Default "bray").

tresh.dist

A scalar informing the threshold value to select eigenvectors based on the amount of variation of each eigenvector. Default is 0.05 (eigenvector with at least 5% of variation). See details.

method.clust

Character indicating the grouping algorithm to be used in cluster analysis. Options avalible are "kmeans" (default) or "ward".

stat.clust

Character to be used in find.clusters indicating the statistic to be computed for each model. Can be "BIC" (default), "AIC" or "WSS".

n.iter.clust

Integer to be used in find.clusters function of adegenet package to indicate the number of iterations to be used in each run of K-means algorithm

criterion.clust

Character string matching "diffNgroup" (default), "min", "goesup", "smoothNgoesup", or "goodfit", indicating the criterion for automatic selection of the optimal number of clusters. See criterion argument in find.clusters for an explanation of these procedures.

Value

A list of length four containing:

PCPS_vectors A matrix with PCPS vectors

cluster_evoregion A vector indicating the region in which each assemblage was classified

Details

evoregions performs biogeographical regionalization analysis, differently from other methods, evoregion uses a phylogenetic turnover metric based on fuzzy sets, therefore accounting for characteristics of evolutionary history, e.g tree imbalance, that is not accounted by other metrics of phylogenetic turnover.

See also

Examples

if (FALSE) {
data(akodon_sites) # occurrence data 
akodon_pa <- akodon_sites %>% 
    dplyr::select(-LONG, -LAT)
data(akodon_newick)
spp_in_tree <- names(akodon_pa) %in% akodon_newick$tip.label
akodon_pa_tree <- akodon_pa[, spp_in_tree]
data(akodon_newick) # phylogenetic tree
regions <- evoregions(comm = akodon_pa_tree, phy = akodon_newick)
site_region <- regions$cluster_evoregions # classification of each community in regions
}